Move REPL loop into its own module
